+++ CHANGES.md
@@ -41,6 +41,7 @@ Bug fixes:
 * VIM-259 Move caret to the line beginning after `==`
 * VIM-246 Fixed `{count}==` formatting
 * VIM-287 Fixed insert new line before and after folds
+* VIM-139 Focus on current esarch and use modal confirmation for `:s///gc`
 
 
 0.44, 2015-11-02